backend/requirements.txt
asgiref==3.7.2
Django==3.2.20
django-filter==23.2
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
